translation/requirements.txt
python-dotenv
langchain
langchain-google-genai
chromadb
PyPDF2
pytesseract
pdf2image
Pillow
sentence-transformers

# Core dependencies
flask==2.3.3
twilio==8.10.0
python-dotenv==1.0.0

# Your existing dependencies
langchain==0.0.350
langchain-google-genai==0.0.6
chromadb==0.4.18

# Additional useful packages
gunicorn==21.2.0  # For production deployment
redis==5.0.1      # For session storage (optional)

#translator related
torch>=1.9.0
transformers>=4.20.0
langdetect>=1.0.9
accelerate>=0.20.0
sentencepiece>=0.1.99
